Patent Payouts for each stage
This commit adds sheet to calculate expenses for patent filing. Each stage has been described in detail along with corresponding expenses for the particular stage.

Note: Professional fee mentioned in the sheet is only for reference. Please add the professional fee as adviced by your IP Lawyer in order to have complete information on the exact payouts for each stage of patent filing. Professional fees is negotiable and differs from lawyer to lawyer.